Avoid forcing the calculation of offsets for pointers.
Exercise #3 needs to put pointers on the stack. By using the same ISA of
machine #2, ithis forces the use of pointers of smaller sizes, which
forces the calculation if real pointers through a base value.

This could still be valuable for one exercise, but I don't think it's
good to have next exercises needing the same care.

So either make #4 change it's ISA to a proper pointer size -- which
makes #4 even more complex -- or make #3 change it's ISA and sacrifice
this base+offset exercise for now.